symm_matrix: option to select set of bands
Added a new argument `bands=slice(None)` to the routine `symm_matrix(...)`, which informs which set of bands will be used in the matrix elements calculation.

The restriction to the selected bands is done at the line `WF = WF[bands]`, and the default value `bands=slice(None)` assure compatibility with previous versions.

Notice that if the code is called without a slice (default parameter), WF remains the original, while if a slice is chosen, a local reduced copy of WF is created. Therefore, the default `bands=slice(None)` also assures that no extra memory is used if the code is called without a slice.