error message suggests shiftUC and refUC
If kname is informed, but kpoint does not match the ones
on the tables, it could be because the correct k point is
related to the informed one by a spacegroup operation.

The new code searches for possible refUC and shiftUC to
fix this discrepancy and presents the alternatives as
a report added to the RuntimeError.